Creo modulo de control de sistema, probado en hw
